+++ src/api/usersApi.js
@@ -3,6 +3,11 @@ import { axiosInstance } from './axios-instance';
 const usersApi = {
   getUsersData: async ({ params }) => await axiosInstance.get('/users', { params }),
 
+  getUserDataByID: async (userId) => {
+    const { data } = await axiosInstance.get(`/users/${userId}`);
+    return data;
+  },
+
   getUserSettingData: async () => await axiosInstance.get('/userSetting'),
 
   getSearchData: async ({ params }) => await axiosInstance.get('/users', { params }),
