+++ src/apis/accountAPI.js
@@ -1,18 +1,18 @@
-import { axiosInstance } from './axios-instance';
+// import { axiosInstance } from './axios-instance';
 
-const accountsAPI = {
-  getAllAccount: async () => {
-    const { data } = await axiosInstance.get('/accounts');
-    return data;
-  },
-  getOneAccountById: async ({ id }) => {
-    const { data } = await axiosInstance.get(`/accounts/${id}`);
-    return data;
-  },
-  updateAccountsById: async ({ id }) => {
-    const { data } = await axiosInstance.patch(`/accounts/${id}`);
-    return data;
-  },
-};
+// const accountsAPI = {
+//   getAllAccount: async () => {
+//     const { data } = await axiosInstance.get('/accounts');
+//     return data;
+//   },
+//   getOneAccountById: async ({ id }) => {
+//     const { data } = await axiosInstance.get(`/accounts/${id}`);
+//     return data;
+//   },
+//   updateAccountsById: async ({ id }) => {
+//     const { data } = await axiosInstance.patch(`/accounts/${id}`);
+//     return data;
+//   },
+// };
 
-export default accountsAPI;
+// export default accountsAPI;
