+++ src/pages/Login/AuthRoute.jsx
@@ -1,8 +1,8 @@
-import { Navigate } from "react-router-dom";
+import { Navigate } from 'react-router-dom';
 import Dashboard from 'components/Dashboard';
 
 const AuthRoute = () => {
-  const auth = Boolean(localStorage.getItem("accessToken"));
+  const auth = Boolean(localStorage.getItem('accessToken'));
 
   return auth ? <Dashboard /> : <Navigate to="/login" />;
 };
